refactor(frontend): move addHexGrid into utlis
